# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/formulas/analyze": {
      "post": {
        "operationId": "FormulaController_analyze",
        "summary": "Analyze formula and resolve external dependencies",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeFormulaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormulaAnalysisResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Formulas"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/reference/regions": {
      "get": {
        "operationId": "RegionsController_findAll",
        "summary": "Get all regions",
        "parameters": [
          {
            "name": "countryId",
            "required": false,
            "in": "query",
            "description": "Filter by country ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "federalDistrictId",
            "required": false,
            "in": "query",
            "description": "Filter by federal district ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "regionTypeId",
            "required": false,
            "in": "query",
            "description": "Filter by region type ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filter by active status",
            "schema": {
              "default": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns list of regions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionsListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Regions"
        ]
      }
    },
    "/reference/regions/{id}": {
      "get": {
        "operationId": "RegionsController_findById",
        "summary": "Get region by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns region",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Region not found"
          }
        },
        "tags": [
          "Regions"
        ]
      }
    },
    "/reference/regions/by-district/{districtId}": {
      "get": {
        "operationId": "RegionsController_findByDistrict",
        "summary": "Get regions by federal district",
        "parameters": [
          {
            "name": "districtId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns list of regions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegionResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Regions"
        ]
      }
    },
    "/reference/regions/by-country/{countryId}": {
      "get": {
        "operationId": "RegionsController_findByCountry",
        "summary": "Get regions by country",
        "parameters": [
          {
            "name": "countryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns list of regions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegionResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Regions"
        ]
      }
    },
...
    "/reference/sports/{id}": {
      "get": {
        "operationId": "SportsController_findById",
        "summary": "Get sport by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns sport",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SportResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Sport not found"
          }
        },
        "tags": [
          "Sports"
        ]
      }
    },
    "/reference/sports/{id}/disciplines": {
      "get": {
        "operationId": "SportsController_findDisciplines",
        "summary": "Get disciplines for a sport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns list of disciplines"
          }
        },
        "tags": [
          "Sports"
        ]
      }
    },
...
    "/reference/seasons/{id}": {
      "patch": {
        "operationId": "SeasonsController_update",
        "summary": "Update an existing season",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Season updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Seasons"
        ]
      },
      "delete": {
        "operationId": "SeasonsController_delete",
        "summary": "Delete a season",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Season deleted"
          }
        },
        "tags": [
          "Seasons"
        ]
      },
      "get": {
        "operationId": "SeasonsController_findById",
        "summary": "Get season by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns season",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Season not found"
          }
        },
        "tags": [
          "Seasons"
        ]
      }
    },
    "/reference/seasons/generate": {
      "post": {
        "operationId": "SeasonsController_generate",
        "summary": "Autogenerate seasons based on logic",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateSeasonsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Seasons generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SeasonResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Seasons"
        ]
      }
    },
    "/reference/seasons/current": {
      "get": {
        "operationId": "SeasonsController_findCurrent",
        "summary": "Get current season",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns current season",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "No current season found"
          }
        },
        "tags": [
          "Seasons"
        ]
      }
    },
...
  },
  "info": {
    "title": "Elbruso API",
    "description": "BI System REST API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AnalyzeFormulaDto": {
        "type": "object",
        "properties": {
          "formula": {
            "type": "string",
            "description": "Formula to analyze",
            "example": "=SUM([Revenue]!A1:A10)"
          },
          "workspaceId": {
            "type": "string",
            "description": "Current workspace ID"
          }
        },
        "required": [
          "formula",
          "workspaceId"
        ]
      },
      "ExternalRefResponseDto": {
        "type": "object",
        "properties": {
          "fullReference": {
            "type": "string",
            "description": "Full reference string",
            "example": "[Revenue]!A1:A10"
          },
          "workspaceId": {
            "type": "string",
            "description": "Target workspace ID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "tableId": {
            "type": "string",
            "description": "Target table ID",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          },
          "tableName": {
            "type": "string",
            "description": "Target table name",
            "example": "Revenue"
          },
          "range": {
            "type": "string",
            "description": "Cell range",
            "example": "A1:A10"
          },
          "hasAccess": {
            "type": "boolean",
            "description": "Whether user has access to the referenced table",
            "example": true
          }
        },
        "required": [
          "fullReference",
          "tableId",
          "tableName",
          "range",
          "hasAccess"
        ]
      },
      "FormulaAnalysisResponseDto": {
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean",
            "description": "Whether formula is syntactically valid",
            "example": true
          },
          "error": {
            "type": "string",
            "description": "Error message if formula is invalid",
            "example": "Unexpected token"
          },
          "externalDependencies": {
            "description": "List of external dependencies",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExternalRefResponseDto"
            }
          }
        },
        "required": [
          "valid",
          "externalDependencies"
        ]
      },
      "RegionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "Region ID"
          },
          "code": {
            "type": "string",
            "example": "moscow",
            "description": "Region code"
          },
          "name": {
            "type": "string",
            "example": "Москва",
            "description": "Region name in Russian"
          },
          "countryId": {
            "type": "number",
            "example": 1,
            "description": "Country ID",
            "nullable": true
          },
          "districtId": {
            "type": "number",
            "example": 1,
            "description": "Federal district ID",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "Whether region is active"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "countryId",
          "districtId",
          "isActive"
        ]
      },
      "RegionsListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "List of regions",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RegionResponseDto"
            }
          },
          "total": {
            "type": "number",
            "example": 100,
            "description": "Total number of regions"
          }
        },
        "required": [
          "items",
          "total"
        ]
      },
      "SportResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "Sport ID"
          },
          "name": {
            "type": "string",
            "example": "Футбол",
            "description": "Sport name in Russian"
          },
          "olympicCategoryId": {
            "type": "number",
            "example": 1,
            "description": "Olympic category ID",
            "nullable": true
          },
          "sportTypeId": {
            "type": "number",
            "example": 1,
            "description": "Sport type ID",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "Whether sport is active"
          }
        },
        "required": [
          "id",
          "name",
          "olympicCategoryId",
          "sportTypeId",
          "isActive"
        ]
      },
...
      "SportReferenceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "Sport ID"
          },
          "name": {
            "type": "string",
            "example": "Football",
            "description": "Sport name"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "SeasonResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "Season ID"
          },
          "code": {
            "type": "string",
            "example": "CY_2024_25",
            "description": "Season code"
          },
          "name": {
            "type": "string",
            "example": "Сезон 2024/2025",
            "description": "Season name in Russian"
          },
          "seasonYear": {
            "type": "number",
            "example": 2024,
            "description": "Season year",
            "nullable": true
          },
          "seasonType": {
            "type": "string",
            "example": "cross_year",
            "description": "Season type",
            "nullable": true
          },
          "startDate": {
            "format": "date-time",
            "type": "string",
            "example": "2024-09-01",
            "description": "Season start date"
          },
          "endDate": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-31",
            "description": "Season end date"
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "Whether season is active"
          },
          "sports": {
            "description": "Sports associated with season",
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SportReferenceDto"
            }
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "seasonYear",
          "seasonType",
          "startDate",
          "endDate",
          "isActive",
          "sports"
        ]
      },
...
      "GenerateSeasonsDto": {
        "type": "object",
        "properties": {
          "startYear": {
            "type": "number",
            "description": "Start year for generation",
            "example": 2024
          },
          "endYear": {
            "type": "number",
            "description": "End year for generation",
            "example": 2030
          },
          "sportId": {
            "type": "number",
            "description": "Specific sport ID (optional)"
          }
        },
        "required": [
          "startYear",
          "endYear"
        ]
      },
...
    }
  }
}